requirements.txt
textblob
azure-functions
